fix conftest.py to actually use astropy machinery like it should be
