+++ astroplan/conftest.py
@@ -4,6 +4,9 @@
 
 from astropy.tests.pytest_plugins import *
 
+# also save a copy of the astropy hooks so we can use them below when overriding
+from astropy.tests import pytest_plugins as astropy_pytest_plugins
+
 ## Uncomment the following line to treat all DeprecationWarnings as
 ## exceptions
 enable_deprecations_as_exceptions()
@@ -18,9 +21,11 @@ except NameError:  # needed to support Astropy < 1.0
     pass
 
 
-
-
 def pytest_configure(config):
+    if hasattr(astropy_pytest_plugins, 'pytest_configure'):
+        # sure ought to be true right now, but always possible it will change in
+        # future versions of astropy
+        astropy_pytest_plugins.pytest_configure(config)
     try:
         import matplotlib
         import nose  # needed for the matplotlib testing tools
@@ -33,6 +38,13 @@ def pytest_configure(config):
             config.option.mpl_baseline_path = 'astroplan/plots/tests/baseline_images'
 
 def pytest_runtest_setup(item):
+    from .utils import _mock_remote_data, _unmock_remote_data
+
+    if hasattr(astropy_pytest_plugins, 'pytest_runtest_setup'):
+        # sure ought to be true right now, but always possible it will change in
+        # future versions of astropy
+        astropy_pytest_plugins.pytest_runtest_setup(item)
+
     # Make appropriate substitutions to mock internet querying methods
     # within the tests
     if item.get_marker('remote_data'):
@@ -41,22 +53,3 @@ def pytest_runtest_setup(item):
     else:
         # a no-op if the last one was not remote-data
         _mock_remote_data()
-
-
-def _mock_remote_data():
-    # Mock FixedTarget.from_name class method for tests without remote data
-    from .target import FixedTarget
-
-    if not hasattr(FixedTarget, '_real_from_name'):
-        FixedTarget._real_from_name = FixedTarget.from_name
-        FixedTarget.from_name = FixedTarget._from_name_mock
-    #otherwise already mocked
-
-def _unmock_remote_data():
-    # undo _mock_remote_data
-    from .target import FixedTarget
-
-    if hasattr(FixedTarget, '_real_from_name'):
-        FixedTarget.from_name = FixedTarget._real_from_name
-        del FixedTarget._real_from_name
-    #otherwise assume it's already correct
